Adds varnish as cache type; can enable/disable from system/cache management, refresh varnish cache

+++ code/Varnish/Helper/Cacheable.php
@@ -3,6 +3,11 @@
 class Magneto_Varnish_Helper_Cacheable extends Mage_Core_Helper_Abstract
 {
 
+    /**
+     * Retrieves current cookie.
+     * 
+     * @return Mage_Core_Model_Cookie
+     */
     public function getCookie()
     {
         return Mage::app()->getCookie();
@@ -24,7 +29,10 @@ class Magneto_Varnish_Helper_Cacheable extends Mage_Core_Helper_Abstract
 
     public function turnOnVarnishCache()
     {
-        $this->getCookie()->delete('nocache');
+        if ($this->getCookie()->get('nocache'))
+        {
+            $this->getCookie()->delete('nocache');
+        }
     }
 
     public function quoteHasItems()
